+++ client/src/app/protected/home/layout.tsx
