+++ client/src/app/protected/home/user/[id]/GroupsLayout.tsx
