+++ client/src/app/protected/home/user/[id]/UserBanner.tsx
