+++ client/src/app/protected/home/user/[id]/page.tsx
